Move the vm_transform action to a mixin
